requirements.txt
pycoin==0.70
python-bitcoinlib==0.5.0
requests==2.9.1
mock==2.0.0
glob2==0.4.1
configargparse==0.10.0
tox==2.3.1
merkle-proofs>=0.0.6
cert-schema>=1.2.19
pyld>=0.7.1